chore(discord): merge request functions into a single file
